# protos/s3_service.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "s3_service.proto",
    "version": "version not set"
  },
...
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/v1/patient/images/upload": {
      "post": {
        "summary": "UploadPatientImage Upload a image of patient",
        "operationId": "PatientImageService_UploadPatientImage",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/s3_serviceUploadPatientImageResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/s3_serviceUploadPatientImageRequest"
            }
          }
        ],
        "tags": [
          "PatientImageService"
        ]
      }
    },
...
  },
  "definitions": {
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
...
    "s3_serviceError": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string"
        },
        "errorRetryStatus": {
          "$ref": "#/definitions/s3_serviceErrorRetryStatus"
        }
      },
      "title": "Error  global error object for api'sresponse"
    },
    "s3_serviceErrorRetryStatus": {
      "type": "string",
      "enum": [
        "NONE",
        "RETRY",
        "NO_RETRY",
        "WAIT_AND_RETRY"
      ],
      "default": "NONE",
      "title": "- NONE: none so user to decide to take further action\n - RETRY: Request can be re-tried\n - NO_RETRY: Request cannot be re-tried\n - WAIT_AND_RETRY: Wait and retry Request"
    },
...
    "s3_serviceImageUI": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "tags": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
...
    "s3_servicePatientImage": {
      "type": "object",
      "properties": {
        "patientId": {
          "type": "string"
        },
        "image": {
          "$ref": "#/definitions/s3_serviceImageUI"
        }
      }
    },
...
    "s3_serviceRequestResult": {
      "type": "string",
      "enum": [
        "UNKNOWN_RESULT",
        "REJECTED",
        "ACCEPTED"
      ],
      "default": "UNKNOWN_RESULT",
      "title": "- REJECTED: Request has been rejected\n - ACCEPTED: Request has been accepted for processing"
    },
    "s3_serviceResult": {
      "type": "object",
      "properties": {
        "requestResult": {
          "$ref": "#/definitions/s3_serviceRequestResult"
        },
        "error": {
          "$ref": "#/definitions/s3_serviceError",
          "description": "Message is a textual description of reason for rejection, and can be safely passed on to operator through logs\nor UI."
        }
      }
    },
...
    "s3_serviceTags": {
      "type": "object",
      "properties": {
        "tag": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
    "s3_serviceUploadPatientImageRequest": {
      "type": "object",
      "properties": {
        "patientImage": {
          "$ref": "#/definitions/s3_servicePatientImage"
        },
        "tags": {
          "$ref": "#/definitions/s3_serviceTags"
        },
        "filePath": {
          "type": "string"
        }
      },
      "title": "CreateOrUpdateBlobRequest - Create or Update a blob in remote storage"
    },
    "s3_serviceUploadPatientImageResponse": {
      "type": "object",
      "properties": {
        "result": {
          "$ref": "#/definitions/s3_serviceResult"
        }
      },
      "title": "CreateOrUpdateBlobResponse Async job id that needs to tracked for success or failure"
    }
  }
}